Backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-redis==5.4.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
gunicorn==23.0.0
python-dotenv==1.0.1
django-celery-beat==2.7.0
